Finish class items.  Prototype data file slurp

+++ Keto_Cta/Visit.cs
@@ -15,7 +15,7 @@ public record Visit
     /// <param name="ncpv">The Non Calcified Plaque Volume.</param>
     /// <param name="tcpv">The Total calcified plaque volume</param>
     /// <param name="pav">The Percent Atheroma Volume</param>
-    public Visit(string id, DateTime? visitDate, double tps, double cac, double ncpv, double tcpv, double pav)
+    public Visit(string id, DateTime? visitDate, int tps, int cac, double ncpv, double tcpv, double pav)
     {
         Id = id;
         VisitDate = visitDate;
@@ -28,8 +28,8 @@ public record Visit
 
     public string Id { get; init; }
     public DateTime? VisitDate { get; init; }
-    public double Tps { get; init; }
-    public double Cac { get; init; }
+    public int Tps { get; init; }
+    public int Cac { get; init; }
     public double Ncpv { get; init; }
     public double Tcpv { get; init; }
     public double Pav { get; init; }
